Small 'fix' for the crash in the Maxima backend.
The problem was (is) that an expression could be deleted when it was
still running, resulting in an dangling pointer being used when maxima
sent the result.
The maxima backend can still crash with a similiar problem, and it is
also possible to make the backend give up on restarting maxima when one
tries to repeatedly re-evaluate an expression in quick succession.
I currently don't see a good solution except for disallowing immediate
reevaluation.