start rewriting the maxima backend
start initial work on a new maxima backend. It makes better use of the
hooks provided by maxima to structure its output and make parsing easier.
It also drops the need for a second maxima process to do the conversion
of output to latex code. Instead the display2d function of maxima is
modified to also deliver the TeX-ified version of the output.
Setting up the maxima session and basic interaction is already working.