better deal with expressions containing comments
we have to detect if an expression only contains of a single comment,
as it otherwise leads to a maxima error when evaluated. This commit rewrites
this detection to better support multiple comments.
Also set the Expression status to done if the expression gets dropped
because its just a comment.