force octave to use gnuplot as plotting backend
In the latest releases of octave it stopped using gnuplot as the default plotting backend,
but the new plotting backend doesn't support to generate a file without showing the plotter window.
So to retain this the cantor feature of being able to show the plot output in the cantor window it's needed to set the graphics_toolkit explicitly.

Patch by Maximiliano Curia, thanks.

REVIEW: 116630