by default use octave-cli rather than octave
'octave' can randomly decide to load its GUI rather than the CLI depending
on environmental factors. We cannot stdin control the GUI, so there is no
point in introducing this instability but instead call octave-cli directly
since it is the only possibly operation mode we support.

(NB there also is a --no-gui switch which for me doesn't appear to work for
 some reason, so octave-cli directly seems the most reliable approach to
 this)

REVIEW: 125646