Make Cantor’s LaTeX renderer friendlier with dark themes
Summary:
When the user selects a dark theme,  the background of the windows (including Cantor’s ones) is dark whereas the text printed in them is white.
It works pretty well in Cantor, expect when the LaTeX renderer is enabled in which case the formulas rendered by LaTeX are still in black on white background. This creates unsightly frames as it can be seen of the following screenshot:
 {F3705639}

This patch makes Cantor pass the currently active background and foreground colors to LaTeX so that LaTeX rendered outputs are looking much more integrated. Please see the following screenshot showing the effect of this patch:
{F3705647}

Reviewers: #cantor, asemke, filipesaraiva

Reviewed By: #cantor, asemke, filipesaraiva

Subscribers: filipesaraiva, asemke, #cantor

Tags: #cantor

Differential Revision: https://phabricator.kde.org/D5457