Added an expression queue to the base session class Session.
At the moment different backends/sessions handle such a queue on their own or don't handle at all. This commit is the first step to unify the different implementations, to provide a standard way to handle multipe expressons and to also reduce the amount of code.
Maxima and R are moved now to the queue in the base class. Remaining backends will be ported/completed step by step in the next commits.