Add support for correct save/load in LatexEntry, when LibSpectre don't installed.
Summary:
In this commit we save rendered image to .cws, in addition to a .eps file. So, if we haven't LibSpectre and can't work with .eps, we load rendered image and show it.
This change allow share worksheet with rendered LatexEntry from user with LibSpectre to user without LibSpectre, who could make changes, and send the worksheet back to first user (and first user don't need rerender LatexEntry and stiill have rendered the entry).

Reviewers: asemke

Reviewed By: asemke

Subscribers: filipesaraiva, asemke, kde-edu

Tags: #kde_edu, #cantor

Differential Revision: https://phabricator.kde.org/D17303