Allow to enable/disable the LaTeX typesetting before the session login.
Reasoning:
In the past the login was done together with the creation of the worksheet and the latex typesetting was set during this step.
After an optimization that was done in this area, the login was postponed to the point in time when the first expression needs to be evaluated.
With this it's not possible to the user to change the latex typesetting until the first evaluation is done which is wrong. We want to allow the user
to modify this setting before anything is processed so also the results of the first evaluation are obtained in the proper format (with or without latex).