[maxima] properly handle the switch to the Lisp mode.
In Maxima it is possible to use Lisp interpreter directly, the switch into this mode is done via to_lisp() call in Maxima. In this case the promt is different and we need to properly handle it. The result of the evaluations are also put at a different place in the output. Also, in the Lisp mode we don't need to update the variable model of Maxima since it's a different environment.