Don't call updateGraphicPackagesFromSettings() in Octave, Python and Julia sessions.
No need to go through this logic for every expression being evaluated. The option for the graphical packages
was only meant to be used by the Plot2DAssistant, this logic only needs to be evaluated when the 2d assistant is used.
Removing these calls for now since it's causing a lot confusion - e.g. the user wants to evaluate '5+5' in julia and get a very long message about the missing GR package.
The proper implementation will come later.