Show the information about the available internal help system in Maxima and in R in the help panel.
Right now, when openning the panel the user only sees and empty area and it's not clear what to do.
This confuses user searching for the documentation and they don't realize this panel is for the internal help systems.
So, we show on default the relevant excerpts from R's and Maxima's documentations describing how to use the internal help system.