Squash commit of multiple fixes that will go into 22.08.
When saving the results, use the file extension and not the mime-type in the QFileDialog to show the relevant files only.

Open URLs in the "Select Backend"-dialog in the external browser.

BUG: 456650
FIXED-IN: 22.08

Enable highdpi pixmaps

Fixes icon rendering on highdpi screens

[maxima] remove the obsolete quotes in string variable values and properly handle quoted sub-strings in string variables.

Moved the worksheet specific actions from the menu "Settings" to the menu "Worksheet" to make it more explicit these actions are affecting the current worksheet only and don't change the global default settings.

Updated ChangeLog.