[tests] skip the execution of Maxima and Octave tests if the executables are not found.
We don't have any compile time dependencies to Maxima and Octave and communicate with the executables.
So, we need to check at runtime whether the executables are present. This will address the issues with failing
Octave tests on CI since Octave is not available in the CI images yet.

Also, added some debug output in one of Maxima's test trying to understand why it's failing on CI.