Squash commit of multiple fixes that need to go into 23.04.
[tests] make initTestCase() in the base class a slot so it's properly found by QtTest
and skip two tests in Maxima and Python that are failing on CI right now,
will be activated once the underlying issue is fixed.

[tests] skip testInterrupt for python.

[lua] refactored and fixed the parsing of LuaJIT's output.

The current implementation doesn't work at all, the output of LuaJIT doesn't fit to the parsing logic (anymore) that was implemented initially.
The refactored code works and handles also multi-line commands where the tests were deactivated earlier for.
More testing and more investment is still probably required to cover more code syntax in Lua/LuaJIT.
This commit makes the lua backend in Cantor working now first, more improvements will come later.

BUG: 461135
FIXED-IN: 23.03

[maxima] always trim the command string and remove line breaks and not only for plot commands so we get one single response in stdout from Maxima
instead of multiple outputs that we only partially handle. This addresses multiple issues with multi-line commands with $ and ; endings.

BUG: 465277
FIXED-IN: 23.03

Added changes that were forgotten in the previous two commits.

When using cantor in labplot, connectToShell() is called after widget() inDocumentationPanelWidget.
We need to add one more connect in connectToShell() so the actions are also working in labplot.

[octave] evaluate the option "integrate plots" on every expression execution and not only during the login
so the user doesn't need to restart Cantor if the options is being changed.

Previously, this option was evaluate at the very beginning and send to Octave via  set (0, "defaultfigurevisible","off")
during the login. In this commit we switch to figure('visible','on') and figure('visible','off') that are prepended to every expression
containing plot commands.

Also, we set the recommended version of Octave to 7.2 which is available for one year already.

[octave] restored the slot OctaveSession::processError() that was deleted in the previous commit.