[octave] evaluate the option "integrate plots" on every expression execution and not only during the login so the user doesn't need to restart Cantor if the options is being changed.
Previously, this option was evaluate at the very beginning and send to Octave via  set (0, "defaultfigurevisible","off")
during the login. In this commit we switch to figure('visible','on') and figure('visible','off') that are prepended to every expression
containing plot commands.

Also, we set the recommended version of Octave to 7.2 which is available for one year already.