[Julia] removed the dependency on Julia's executable and set the recommended version to the current LTS version 1.6.7.
We don't communicate via CLI with julia's executable and load the interpreter
into Cantor's runtime in the "server" directly and similarly to how it's
done also for Python and R. cantor_juliabackend.so and and cantor_juliaserverjulia
are linked to julia's libraries. We nowhere depend on the executable.
Allowing the user to specify the path to the executable and loading the libraries
from there made some edge cases possible but introduced additional complexity
in the code and such severe limitation as not being able to work with symlinks
in the PATH. So, we remove this complexity and limitations now and the more important
option to work with different packages and version of Python, R and Julia will be
solved in future differently.

BUG: 470381
FIXED-IN: 23.12