Upgrade third-party Discount library from 2.2.6 to 3.0.0e
Original modifications for extracting LaTeX from Markdown
for future processing inside Cantor were also ported
between versions. Ideally, they would need to be upstreamed,
but this is an ongoing and not very fast process, so we
are using a local copy of this library with local
modifications for the time being.