Fix hardcoded library prefix usage by CMake variable
The third-party project was using a hardcoded prefix in
the CMake source code to calculate the compiled library's
full name, but this was not a cross-platform solution and
was causing compilation failures in the Qt5 + Windows
environment. These changes replace the use of this
hardcoded value with the special CMake variable
CMAKE_STATIC_LIBRARY_PREFIX, which provides a proper
cross-platform solution.