(Re-)Enable the embedded documentation, set it to ON on default.
CI: Enable embedded documentation

Even if it isn't enabled (or maybe even especially because of this),
CI coverage of these parts of the code base would be nice.

Fix the embedded documentation Qt6 port

The submodule for Qt Quick has been renamed from QtWebEngine to
QtWebEngineQuick. I'm not sure, but I assume that slightly imprecise
previous name lead to the inclusion of QtWebEngine::initialize(). But I
don't think it's necessary because there are no Quick things used with
cantor.